[#58] Use lookup for files as well, fix version number
Fix version number to 0.6.0-DEV after merge conflict.
Save files in the flight metadata as UInt16 and refactor the roots
lookup dictionary to have fields "roots" and "files".
Include routines from filesystem.jl in datachecks.jl and remove
filesystem.jl.
Optimize file loading for web data. Add exception handling for web data
file reading.
Rename sattypes_new.jl to sattypes.jl and sattypes.jl to
observations.jl.
Include the previously broken test after merge of fix/0.5.4.
