+++ src/flightdata.jl
@@ -3,13 +3,14 @@
 """
     load_volpe(
         paths::Vector{@NamedTuple{root::String,files::Vector{String}}},
-        roots::ds.OrderedDict{String,UInt16},
+        pathdict::ds.OrderedDict{String,ds.OrderedDict},
         altmin::Real=5_000,
         Float::DataType=Float32
     ) -> StructArray{FlightData{Float}}
 
-From a list of `files`, return a `StructArray{FlightData{Float}}` that can
-be saved to the `inventory` field in `FlightSet`.
+From a list of `paths`, return a `StructArray{FlightData{Float}}` that can
+be saved to the `volpe` field in `FlightSet`. The `pathdict` connects file names with integer
+indices used in the metadata of `FlightData` and `FlightSet` to save memory.
 
 When the `StructArray{FlightData{Float}}` is constructed, data can be filtered by a minimum
 altitude threshold in meters of the aircraft data (default: `altmin=5_000`).
@@ -18,7 +19,7 @@ by default `Float32`.
 """
 function load_volpe(
     paths::Vector{@NamedTuple{root::String,files::Vector{String}}},
-    roots::ds.OrderedDict{String,UInt16},
+    pathdict::ds.OrderedDict{String,ds.OrderedDict},
     altmin::Real=5_000,
     Float::DataType=Float32
 )::StructArray{FlightData{Float}}
@@ -58,10 +59,10 @@ function load_volpe(
         for group in groups
             df.nrow(group) ≤ 1 && continue
             track = copy(group) # ℹ avoid problems with views when modifying data
-            df.select!(track, df.Not(:FID))
             flex, useLON = preptrack!(track)
             isempty(flex) || push!(inventory, FlightData{Float}(track, group.FID[1],
-                missing, missing, missing, flex, useLON, 0x01, roots[path.root], file))
+                missing, missing, missing, flex, useLON, 0x01, pathdict["roots"][path.root],
+                pathdict["files"][file]))
         end #loop over flights
         # Monitor progress for progress bar
         pm.next!(prog, showvalues = [(:file,splitext(basename(file))[1])])
@@ -73,20 +74,25 @@ end #function loadVOLPE
 
 
 """
-    load_flightaware(files::String...; Float::DataType=Float32, altmin::Real=5_000)
-        -> StructArray{FlightData{Float}}
+    load_flightaware(
+        paths::Vector{@NamedTuple{root::String,files::Vector{String}}},
+        pathdict::ds.OrderedDict{String,ds.OrderedDict},
+        altmin::Real=5_000,
+        Float::DataType=Float32
+    ) -> StructArray{FlightData{Float}}
 
-From a list of `files`, return a `Vector{FlightTrack}` that can
-be saved to the `archive` field in `FlightSet`.
+From a list of `paths`, return a `StructArray{FlightData{Float}}` that can
+be saved to the `flightaware` field in `FlightSet`. The `pathdict` connects file names with
+integer  indices used in the metadata of `FlightData` and `FlightSet` to save memory.
 
-When the `Vector{FlightTrack{T}}` is constructed, data can be filtered by a minimum
+When the `StructArray{FlightData{Float}}` is constructed, data can be filtered by a minimum
 altitude threshold in meters of the aircraft data (default: `altmin=5_000`).
 Floating point numbers in `FlightTrack` are of the precision set by `Float`,
 by default `Float32`.
 """
 function load_flightaware(
     paths::Vector{@NamedTuple{root::String,files::Vector{String}}},
-    roots::ds.OrderedDict{String,UInt16},
+    pathdict::ds.OrderedDict{String,ds.OrderedDict},
     altmin::Real=5_000,
     Float::DataType=Float32
 )::StructArray{FlightData{Float}}
@@ -103,7 +109,6 @@ function load_flightaware(
         flights.alt = ft2m.(flights.alt)
         flights.speed = knot2mps.(flights.speed)
         flights.climb = ftpmin2mps.(flights.climb)
-
         # Filter data: keep if lat/lon not missing and alt is missing or alt >= altmin
         flights = flights[(.!ismissing.(flights.lat)) .& (.!ismissing.(flights.lon)) .&
             (ismissing.(flights.alt) .| (flights.alt .≥ altmin)), :]
@@ -111,22 +116,18 @@ function load_flightaware(
         # Group by flight ID
         groups = df.groupby(flights, :dbID)
         sizehint!(archive, length(archive) + length(groups))
-
         # Process each flight
         for group in groups
             df.nrow(group) ≤ 1 && continue
-
             # Copy to DataFrame and select track columns
             track = df.select(copy(group), :time, :lat, :lon, :alt, :speed, :climb, :heading)
-
             # Determine predominant flight direction, inflection points, and remove duplicates
             flex, useLON = preptrack!(track)
-
             # Save the FlightTrack in the archive vector
             isempty(flex) || push!(archive, FlightData{Float}(track, group.dbID[1],
                 group.flightID[1], group.type[1],
                 (orig=group.orig[1], dest=group.dest[1]), flex, useLON,
-                0x02, roots[path.root], file))
+                0x02, pathdict["roots"][path.root], pathdict["files"][file]))
         end #loop over flights
 
         # Monitor progress for progress bar
@@ -141,14 +142,15 @@ end #function load_flightaware
 """
     load_webdata(
         paths::Vector{@NamedTuple{root::String,files::Vector{String}}},
-        roots::ds.OrderedDict{String,UInt16},
+        pathdict::ds.OrderedDict{String,ds.OrderedDict},
         altmin::Real=5_000,
         Float::DataType=Float32,
         delim::Union{Nothing,Char,String}=nothing
     ) -> StructArray{FlightData{Float}}
 
-From a list of `files`, return a `Vector{FlightTrack}` that can
-be saved to the `onlineData` field in `FlightSet`.
+From a list of `paths`, return a `StructArray{FlightData{Float}}` that can
+be saved to the `webdata` field in `FlightSet`. The `pathdict` connects file names with
+integer  indices used in the metadata of `FlightData` and `FlightSet` to save memory.
 
 The `delim`iter of the data in the input file can be specified by a string or character.
 Default is `nothing`, which means auto-detection of the delimiter is used.
@@ -160,7 +162,7 @@ by default `Float32`.
 """
 function load_webdata(
     paths::Vector{@NamedTuple{root::String,files::Vector{String}}},
-    roots::ds.OrderedDict{String,UInt16},
+    pathdict::ds.OrderedDict{String,ds.OrderedDict},
     altmin::Real=5_000,
     Float::DataType=Float32,
     delim::Union{Nothing,Char,String}=nothing
@@ -168,14 +170,22 @@ function load_webdata(
     # Initialise inventory file array
     archive = StructArray{FlightData{Float}}(undef, 0)
     isempty(paths) && return archive
+    nlength = sum(length.(getfield.(paths, :files)))
+    sizehint!(archive, nlength)
     # Loop over files with online data
-    prog = pm.Progress(sum(length.(paths.files)), desc = "load online data...")
+    prog = pm.Progress(nlength, desc = "load web data...")
     for path in paths, file in path.files
         # Read flight data
-        flight = CSV.read(file, DataFrame, delim=delim, ignoreemptylines=true, normalizenames=true,
-        silencewarnings=true, ntasks=1, copycols=false, types=Dict(:Latitude => Float,
-        :Longitude => Float, :feet => String, :kts => Float, :Course => String,
-        :Rate => String), drop = ["mph", "Reporting_Facility"], stringtype=String)
+        flight = try CSV.read(joinpath(path.root, file), DataFrame, delim=delim, ignoreemptyrows=true,
+            normalizenames=true, silencewarnings=true, ntasks=1, copycols=false,
+            types=Dict(:Latitude => Float, :Longitude => Float, :feet => String, :kts => Float,
+            :Course => String, :Rate => String), drop = ["mph", "Reporting_Facility"], stringtype=String)
+        catch err
+            println()
+            println()
+            @error "Error reading file. Try to specify column delimiter. Data skipped." file exception=(err, catch_backtrace())
+            continue
+        end
         # Convert knots to m/s
         flight.kts = knot2mps.(flight.kts)
         if length(names(flight)) ≠ 7 || names(flight)[2:7] ≠
@@ -201,56 +211,67 @@ function load_webdata(
 
         # Set to 2 days prior to allow corrections for timezone diffences in the next step
         date += Dates.Day(2)
-        ### Convert times to datetime and extract heading and climbing rate as Int
-        # Initialise time vector
+        ### Convert times to datetime and extract heading and climbing rate
+        # Initialise vectors
+        lat = Float[]
+        lon = Float[]
         flighttime = ZonedDateTime[]
-        # Initialise vectors for altitude, heading and climb to convert from strings to Int
         altitude = Union{Missing,Float}[]
         heading = Union{Missing,Int}[]
         climbingrate = Union{Missing,Float}[]
-        # Loop over times
-        for i=length(flight[!,1]):-1:1
-        alt = try ft2m(parse(Float, join([n for n in flight.alt[i]
-            if isnumeric(n) || n == '.'])))
-        catch; missing;  end
-        climb = try ftpmin2mps(parse(Float, join([n for n in flight.climb[i]
-            if isnumeric(n) || n == '.' || n == '-'])))
-        catch; missing;  end
-        head = try parse(Int, join([n for n in flight.heading[i]
-            if isnumeric(n) || n == '.']))
-        catch; missing;  end
-        if length(flight.time[i]) ≠ 15 || ismissing(flight.lat[i]) ||
-            ismissing(flight.lon[i]) || (!ismissing(alt) && alt < altmin)
-            delete!(flight, i)
-            continue
-        end
-        # Derive date from day of week and filename
-        while Dates.dayabbr(date) ≠ flight.time[i][1:3]
-            date -= Dates.Day(1)
-        end
-        # Derive time from time string
-        t = Time(flight.time[i][5:end], "I:M:S p")
-        # Save data that needed tweaking of current time step
-        pushfirst!(flighttime, ZonedDateTime(DateTime(date, t), timezone))
-        pushfirst!(altitude, alt); pushfirst!(climbingrate, climb); pushfirst!(heading, head)
+        speed = Union{Missing,Float}[]
+        # Loop over times (backwards to preserve day matching logic)
+        for i = length(flight[!,1]):-1:1
+            alt = try ft2m(parse(Float, join([n for n in flight.alt[i]
+                if isnumeric(n) || n == '.'])))
+            catch; missing; end
+            climb = try ftpmin2mps(parse(Float, join([n for n in flight.climb[i]
+                if isnumeric(n) || n == '.' || n == '-'])))
+            catch; missing; end
+            head = try parse(Int, join([n for n in flight.heading[i]
+                if isnumeric(n) || n == '.']))
+            catch; missing; end
+            if length(flight.time[i]) ≠ 15 || ismissing(flight.lat[i]) ||
+                ismissing(flight.lon[i]) || (!ismissing(alt) && alt < altmin)
+                continue
+            end
+            # Derive date from day of week and filename
+            while Dates.dayabbr(date) ≠ flight.time[i][1:3]
+                date -= Dates.Day(1)
+            end
+            # Derive time from time string
+            t = Time(flight.time[i][5:end], "I:M:S p")
+            # Save data (reverse order for now)
+            push!(flighttime, ZonedDateTime(DateTime(date, t), timezone))
+            push!(altitude, ismissing(alt) ? missing : Float(alt))
+            push!(climbingrate, ismissing(climb) ? missing : Float(climb))
+            push!(heading, head)
+            push!(lat, Float(flight.lat[i]))
+            push!(lon, Float(flight.lon[i]))
+            push!(speed, ismissing(flight.speed[i]) ? missing : Float(flight.speed[i]))
         end #loop of flight
 
         # Skip data with all data points below the altitude threshold or missing
         isempty(altitude) && continue
-        # Save revised data to DataFrame
-        flight.time = flighttime
-        flight.alt = altitude
-        flight.climb = climbingrate
-        flight.heading = heading
-        # Convert lat/lon from type Union{Missing,Float64} to Float64 to be processed by findflex
-        flight.lat = float.(flight.lat); flight.lon = float.(flight.lon)
+        # Restore original order
+        reverse!(flighttime)
+        reverse!(altitude)
+        reverse!(climbingrate)
+        reverse!(heading)
+        reverse!(lat)
+        reverse!(lon)
+        reverse!(speed)
+        # Build track DataFrame
+        track = DataFrame(time = flighttime, lat = lat, lon = lon, alt = altitude,
+            speed = speed, climb = climbingrate, heading = heading)
 
         # Determine predominant flight direction, inflection points, and remove duplicate entries
-        flex, useLON = preptrack!(flight)
+        flex, useLON = preptrack!(track)
 
         # Save data as FlightTrack
-        isempty(flex) || push!(archive, FlightData{Float}(flight, replace(filename, "_" => "/"),
-        flightID, missing, (orig=orig, dest=dest), flex, useLON, 0x03, roots[path.root], file))
+        isempty(flex) || push!(archive, FlightData{Float}(track, replace(filename, "_" => "/"),
+            flightID, missing, (orig=orig, dest=dest), flex, useLON, 0x03,
+            pathdict["roots"][path.root], pathdict["files"][file]))
         # Monitor progress for progress bar
         pm.next!(prog)
     end #loop over files
