+++ src/sattypes.jl
@@ -1,391 +1,318 @@
-## ObservationSet
+## structs for satellite data
 
 """
-# struct CLay
+# struct SecondaryMetadata{T<:AbstractFloat} <: SecondarySet{T}
 
-CALIOP cloud layer `data` stored in a `DataFrame` with columns:
-- `time::Vector{DateTime}` (time index)
-- `lat::Vector{AbstractFloat}` (latitude position of current time index)
-- `lon::Vector{AbstractFloat}` (longitude position of current time index)
-- `layer::Vector{NamedTuple{(:top,:base),Tuple{Vector{<:AbstractFloat},Vector{<:AbstractFloat}}}}`
-  (layer top/base heights in meters)
-- `atmos_state::Vector{Vector{Symbol}}` (symbols describing the atmospheric conditions at the intersection)
-- `OD::Vector{<:Vector{<:AbstractFloat}}` (layer optical depth)
-- `IWP::Vector{<:Vector{<:Union{Missing,<:AbstractFloat}}}` (layer ice water path)
-- `Ttop::Vector{<:Vector{<:AbstractFloat}}` (layer top temperature)
-- `Htropo::Vector{<:AbstractFloat}` (tropopause height at current time index)
-- `night::BitVector` (flag for nights (`true`))
-- `averaging::Vector{<:Vector{Int}}` (horizontal averaging in km)
+Immutable struct to hold metadata for `SatSet{T}` with fields
 
-# Instantiation
+- `files::Dict{Int,String}`
+- `type::Symbol`
+- `date::NamedTuple{(:start,:stop),Tuple{DateTime,DateTime}}`
+- `created::Union{DateTime,ZonedDateTime}`
+- `loadtime::Dates.CompoundPeriod`
+- `remarks`
 
-    function CLay{T}(
-      ms::mat.MSession,
-      files::Vector{String},
-      timespan::NamedTuple{(:min,:max), Tuple{DateTime,DateTime}},
-      lidarrange::Tuple{Real,Real}=(15_000,-Inf),
-      altmin::Real=5000
-    ) where T
+## fields
 
-Construct `CLay` from a list of file names (including directories) and a running
-MATLAB session `ms` and save data, if layers are within the bounds
-of `lidarrange` and above flight `altmin` threshold and time is within `timesapn`.
-If `T<:AbstractFloat` is not set, `Float32` will be used as default precision.
+### granules
 
-Or construct `CLay` by directly handing over the `DataFrame` where the names, order,
-and types of each columns are checked and attempted to correct:
+DataFrame with complimentary information on each data file, such as file `name` and `root`,
+`tstart` and `tstop` time, and lat/lon extrema (overall and per hemisphere).
 
-    CLay{T}(data::DataFrame) where T -> struct CLay
+### roots
+
+OrderedDict connecting root paths with indices for more efficient storage in `granules`.
+
+### type
+
+Symbol indicating, whether profile or layer data is stored.
+
+### date
+
+`NamedTuple` with fields `start` and `stop` for start and end time of the monitored
+satellite period.
+
+### created
+
+time of creation of database
+
+### loadtime
+
+time it took to read data files and load it to the struct
+
+### attachments
+
+any additional data or comments that can be attached to the database
+
+
+## Instantiation
+
+`SecondaryMetadata` is constructed automatically, when `SatSet` is instantiated using
+a modified constructor. Any `attachments` can be passed to the metadata with the keyword
+argument `attachments` when constructing `SatSet`.
+
+Additional constructors help to instantiate empty `SecondaryMetadata` and to promote the
+floating point type of an existing `SecondaryMetadata`. The floating point type can be omitted,
+in which case `Float32` will be used by default.
 """
-struct CLay{T} <: ObservationSet{T}
-  data::DataFrame
+struct SecondaryMetadata{T<:AbstractFloat} <: SecondarySet{T}
+    granules::DataFrame
+    roots::ds.OrderedDict{UInt16,String}
+    type::Symbol
+    date::NamedTuple{(:start,:stop),Tuple{DateTime,DateTime}}
+    created::AbstractDateTime
+    loadtime::Dates.CompoundPeriod
+    attachments
 
-  """ Unmodified constructor for `CLay` """
-  function CLay{T}(data::DataFrame) where T
-    # Ensure floats of correct precision
-    convert_floats!(data, T)
-    # Column checks and warnings
-    standardnames = ["time", "lat", "lon", "layer_top", "layer_base", "atmos_state",
-      "OD", "IWP", "Ttop", "Htropo", "night", "averaging"]
-    standardtypes = [Vector{DateTime}, Vector{T}, Vector{T},
-      Vector{<:Vector{<:T}}, Vector{<:Vector{<:T}},
-      Vector{<:Vector{Symbol}}, Vector{<:Vector{<:T}},
-      Vector{<:Vector{<:Union{Missing,<:T}}}, Vector{<:Vector{<:T}},
-      Vector{<:T}, BitVector, Vector{<:Vector{<:Int}}]
-    bounds = (:lat => (-90,90), :lon => (-180,180))
-    checkcols!(data, standardnames, standardtypes, bounds, "CLay")
-    new{T}(data)
-  end #constructor 1 CLay
+    #* Internal constructor with data checks
+    function SecondaryMetadata{T}(
+        granules::DataFrame,
+        roots::ds.OrderedDict{UInt16,String},
+        type::Symbol,
+        date::NamedTuple{(:start,:stop),Tuple{DateTime,DateTime}},
+        created::AbstractDateTime,
+        loadtime::Dates.CompoundPeriod,
+        attachments
+    ) where T<:AbstractFloat
+    # Check that required columns are present in granules DataFrame
+    colnames = ["file", "root", "tstart", "tstop", "latmin", "latmax", "elonmin", "elonmax", "wlonmin", "wlonmax"]
+    coltypes = [Vector{String}, Vector{UInt16}, Vector{DateTime}, Vector{DateTime},
+        Vector{T}, Vector{T}, Vector{T}, Vector{T}, Vector{T}, Vector{T}]
+    essentialcols = collect(1:length(colnames))
+    bounds = (:tstart => (Date(2000), Dates.now()), :tstop => (Date(2000), Dates.now()),
+        :latmin => (-90, 90), :latmax => (-90, 90),
+        :elonmin => (0, 180), :elonmax => (0, 180), :wlonmin => (-180, 0), :wlonmax => (-180, 0))
+    checkcols!(granules, colnames, coltypes, bounds, "SecondaryMetadata"; essentialcols)
+    granule_roots = (sort∘unique)(granules.root)
+    root_keys = roots.keys
+    granule_roots == root_keys ||
+        throw(ArgumentError("values in root column of granules must match keys of the roots OrderedDict\n"*
+            "got: $granule_roots\nvs.: $root_keys"))
+    all(granules.tstart .≤ granules.tstop) ||
+        throw(ArgumentError("tstart must be less than or equal to tstop for all granules"))
+    all(granules.latmin .≤ granules.latmax) ||
+        throw(ArgumentError("latmin must be less than or equal to latmax for all granules"))
+    all(granules.elonmin .≤ granules.elonmax) ||
+        throw(ArgumentError("elonmin must be less than or equal to elonmax for all granules"))
+    all(granules.wlonmin .≤ granules.wlonmax) ||
+        throw(ArgumentError("wlonmin must be less than or equal to wlonmax for all granules"))
+    new{T}(granules, roots, type, date, created, loadtime, attachments)
+  end
+end #struct SecondaryMetadata
 
-  """
-  Modified constructor of `CLay` reading data from hdf `files` using MATLAB session `ms`
-  in the `lidarrange` (top to bottom), if data is above `altmin`.
-  """
-  function CLay{T}(
-    ms::mat.MSession,
-    files::Vector{String},
-    timeindex::Vector{UnitRange{Int}},
-    lidarrange::Tuple{Real,Real}=(15_000,-Inf),
-    altmin::Real=5000,
-    saveobs::Union{String,Bool}="abs"
-  ) where T
-    # Return default empty struct if files are empty
-    (isempty(files) || saveobs === false || isempty(saveobs)) && CLay()
-    # Initialise arrays
-    # essential data
-    utc = Vector{Vector{DateTime}}(undef, length(files))
-    lat = Vector{Vector{T}}(undef, length(files))
-    lon = Vector{Vector{T}}(undef, length(files))
-    # non-essential data
-    LayTop = Vector{Vector{Vector{T}}}(undef, length(files))
-    LayBase = Vector{Vector{Vector{T}}}(undef, length(files))
-    Atmosph = Vector{Vector{Vector{Symbol}}}(undef,length(files))
-    OD = Vector{Vector{Vector{T}}}(undef,length(files))
-    IWP = Vector{Vector{Vector{Union{Missing,T}}}}(undef,length(files))
-    Ttop = Vector{Vector{Vector{T}}}(undef,length(files))
-    Htropo = Vector{Vector{T}}(undef, length(files))
-    night = Vector{BitVector}(undef, length(files))
-    averaging = Vector{Vector{Int}}(undef,length(files))
-    # Loop over files
-    for (i, file) in enumerate(files)
-      # Send file name to MATLAB
-      mat.put_variable(ms, :file, file)
-      # Read hdf file with MATLAB
-      mat.eval_string(ms,
-        """
-        clear t longitude latitude basealt topalt ...
-          FCF FOD IWPath LTT Htropo daynight average
-        try
-          t = hdfread(file, 'Profile_UTC_Time');
-          t = t(:, 2);
-          longitude = hdfread(file, 'Longitude');
-          longitude = longitude(:, 2);
-          latitude = hdfread(file, 'Latitude');
-          latitude = latitude(:, 2);
-          basealt = hdfread(file, 'Layer_Base_Altitude');
-          topalt = hdfread(file, 'Layer_Top_Altitude');
-          FCF = hdfread(file, 'Feature_Classification_Flags');
-          FOD = hdfread(file, 'Feature_Optical_Depth_532');
-          IWPath = hdfread(file, 'Ice_Water_Path');
-          LTT = hdfread(file, 'Layer_Top_Temperature');
-          Htropo = hdfread(file, 'Tropopause_Height');
-          daynight = hdfread(file, 'Day_Night_Flag');
-          average = hdfread(file, 'Horizontal_Averaging');
-        end
+#* Constructor for empty SecondaryMetadata
+SecondaryMetadata{T}() where T = SecondaryMetadata{T}(
+    DataFrame(file = String[], root = UInt16[], tstart = DateTime[], tstop = DateTime[],
+        latmin = T[], latmax = T[], elonmin = T[], elonmax = T[], wlonmin = T[], wlonmax = T[]
+    ),
+    ds.OrderedDict{UInt16,String}(),
+    :undef, (start=Dates.now(), stop=Dates.now()), Dates.now(), Dates.CompoundPeriod(), nothing
+)
 
-        data = struct('time', t, 'lat', latitude, 'lon', longitude, ...
-          'LB', basealt, 'LT', topalt, 'FCF', FCF, 'FOD', FOD, 'IWP', IWPath, ...
-          'LTT', LTT, 'Htropo', Htropo, 'daynight', daynight, 'average', average);
-        """
-      )
+#* Constructor for default Float32 SecondaryMetadata
+SecondaryMetadata(args...) = SecondaryMetadata{Float32}(args...)
 
-      # Get data from MATLAB
-      data = mat.jdict(mat.get_mvariable(ms, :data))
-      # Convert time to UTC DateTime
-      utc[i] = convert_utc.(data["time"])
-      utc[i] = utc[i][timeindex[i]]
-      # Extract lat/lon
-      lat[i], lon[i] = data["lat"][timeindex[i]], data["lon"][timeindex[i]]
-
-      ## Extract non-essential information from satellite files
-      Ltop, Lbase = 1000data["LT"][timeindex[i],:], 1000data["LB"][timeindex[i],:]
-      FCF = data["FCF"][timeindex[i],:]
-      FOD = data["FOD"][timeindex[i],:]
-      IWPath = data["IWP"][timeindex[i],:]
-      LTT = data["LTT"][timeindex[i],:]
-      Htropo[i] = 1000data["Htropo"][timeindex[i]]
-      night[i] = Bool.(data["daynight"][timeindex[i]])
-      averaging[i] = data["average"][timeindex[i]]
-
-      # Loop over data and convert to TrackMatcher format
-      Lt = Vector{Vector{T}}(undef,length(utc[i]))
-      Lb = Vector{Vector{T}}(undef,length(utc[i]))
-      atm = Vector{Vector{Symbol}}(undef,length(utc[i]))
-      optdepth = Vector{Vector{T}}(undef,length(utc[i]))
-      icewater = Vector{Vector{Union{Missing,T}}}(undef,length(utc[i]))
-      toptemp = Vector{Vector{T}}(undef,length(utc[i]))
-      for n = 1:length(utc[i])
-        l = findall((Lbase[n,:] .> 0) .& (Ltop[n,:] .> 0) .& (Lbase[n,:] .< lidarrange[1]) .&
-          (Ltop[n,:] .> lidarrange[2]) .& (Ltop[n,:] .> altmin))
-        Lt[n], Lb[n], atm[n], optdepth[n], toptemp[n], icewater[n] =
-          if isempty(l)
-            T[], T[], Symbol[], T[], T[], T[]
-          else
-            l = findall((Lbase[n,:] .> 0) .& (Ltop[n,:] .> 0) .& (Lbase[n,:] .< lidarrange[1]) .&
-              (Ltop[n,:] .> lidarrange[2]))
-            [Ltop[n, m] for m in l] , [Lbase[n, m] for m in l],
-            [feature_classification(classification(FCF[n,m])...) for m in l],
-            [FOD[n,m] for m in l],
-            [LTT[n,m] for m in l],
-            [IWPath[n,m] == -9999 ? missing : IWPath[n,m] for m in l]
-          end
-      end # loop over time steps in current file
-      LayTop[i], LayBase[i], Atmosph[i], OD[i], IWP[i], Ttop[i] =
-        Lt, Lb, atm, optdepth, icewater, toptemp
-    end #loop over files
-
-    # Construct and standardise data
-    data = DataFrame(time=[utc...;], lat=[lat...;], lon=[lon...;],
-      layer_top=[LayTop...;], layer_base=[LayBase...;], atmos_state=[Atmosph...;],
-      OD=[OD...;], IWP=[IWP...;], Ttop=[Ttop...;], Htropo = [Htropo...;],
-      night = [night...;], averaging = [averaging...;])
-    # Save time, lat/lon arrays in CLay struct
-    new{T}(data)
-  end #constructor 2 CLay
-end #struct CLay
+#* Constructor for floating point type promotion
+SecondaryMetadata{T}(meta::SecondaryMetadata) where T = SecondaryMetadata{T}(
+    DataFrame(file = meta.granules.file, root = meta.granules.root,
+        tstart = meta.granules.tstart, tstop = meta.granules.tstop,
+        latmin = T.(meta.granules.latmin), latmax = T.(meta.granules.latmax),
+        elonmin = T.(meta.granules.elonmin), elonmax = T.(meta.granules.elonmax),
+        wlonmin = T.(meta.granules.wlonmin), wlonmax = T.(meta.granules.wlonmax)
+    ),
+    meta.roots, meta.type, meta.date, meta.created, meta.loadtime, meta.attachments
+)
 
 
 """
-    CLay{T}() where T
+# struct SatData{T<:AbstractFloat} <: SatTrack{T}
 
-External constructor for empty `CLay` struct.
+Satellite track data storing a single granule from a individual satellite data file with fields
+
+- `time::Vector{DateTime}`: UTC time of satellite measurements
+- `lat::Vector{T}`: latitudes of satellite measurements
+- `lon::Vector{T}`: longitudes  of satellite measurements
+
+see also [`SatTrack`](@ref) and [`SatSet`](@ref)
 """
-function CLay{T}() where T
-  data = DataFrame(time = DateTime[], lat = T[], lon = T[],
-  layer_top = Vector{T}[], layer_base = Vector{T}[],
-  atmos_state = Vector{Symbol}[], OD = Vector{T}[], IWP = Vector{T}[],
-  Ttop = Vector{T}[], Htropo = T[], night = BitVector(), averaging = Vector{Int}[])
-  CLay{T}(data)
-end
+struct SatData{T<:AbstractFloat} <: SatTrack{T}
+    time::Vector{DateTime}
+    lat::Vector{T}
+    lon::Vector{T}
 
-"""
-    CLay{T}(clay::CLay) where T
-
-External `CLay` constructor for conversion of floating point precision.
-"""
-function CLay{T}(clay::CLay) where T
-  convert_floats!(clay.data, T)
-  CLay{T}(clay.data)
-end
-
-""" Default CLay constructor for Float32 """
-CLay(args...; kwargs...) = CLay{Float32}(args...; kwargs...)
-
-
-"""
-# struct CPro
-
-CALIOP cloud profile `data` stored in a `DataFrame` with columns:
-- `time::Vector{DateTime}` (current time index)
-- `lat::Vector{AbstractFloat}` (latitude coordinate for current time index)
-- `lon::Vector{AbstractFloat}` (lonitude coordinate for current time index)
-- `atmos_state::Vector{<:Vector{<:Union{Missing,Symbol}}}`
-  (symbols describing the atmospheric conditions for every height level at current time index)
-- `EC532::Vector{<:Vector{<:Union{Missing,AbstractFloat}}}`
-  (extinction coefficient at 532nm at every height level in current time index)
-
-# Instantiation
-
-    function CPro{T}(
-      ms::mat.MSession,
-      files::Vector{String},
-      timespan::NamedTuple{(:min,:max), Tuple{DateTime,DateTime}},
-      lidarprofile::NamedTuple
-    ) where T -> struct CPro
-
-Construct `CPro` from a list of file names (including directories) and a running
-MATLAB session `ms`. CPro data is only stored in the vicinity of intersections for
-the designated `timespan`. Column data is stored height-resolved as defined by the
-`lidarprofile`. If `T<:AbstractFloat` is not set, `Float32` will be used as
-default precision.
-
-Or construct `CPro` by directly handing over the `DataFrame` where the names, order,
-and types of each columns are checked and attempted to correct:
-
-    CPro{T}(data::DataFrame) where T -> struct CPro
-"""
-struct CPro{T} <: ObservationSet{T}
-  data::DataFrame
-
-  """ unmodified constructor """
-  function CPro{T}(data::DataFrame) where T
-    # Ensure floats of correct precision
-    convert_floats!(data, T)
-    # Column checks and warnings
-    standardnames = ["time", "lat", "lon", "atmos_state", "EC532", "Htropo", "temp",
-      "pressure", "rH", "IWC", "deltap", "CADscore", "night"]
-    standardtypes = [Vector{DateTime}, Vector{<:T}, Vector{<:T},
-      Vector{<:Vector{<:Union{Missing,Symbol}}}, Vector{<:Vector{<:Union{Missing,<:T}}},
-      Vector{<:T}, Vector{<:Vector{<:Union{Missing,<:T}}},
-      Vector{<:Vector{<:Union{Missing,<:T}}}, Vector{<:Vector{<:Union{Missing,<:T}}},
-      Vector{<:Vector{<:Union{Missing,<:T}}}, Vector{<:Vector{<:Union{Missing,<:T}}},
-      Vector{<:Vector{<:Union{Missing,Int8}}}, BitVector]
-    bounds = (:lat => (-90,90), :lon => (-180,180), :Htropo => (4000,22_000),
-      :temp => (-120,60), :pressure => (1,1086), :rH => (0,1.5), :IWC => (0,0.54),
-      :deltap => (0,1), :CADscore => (-101,106))
-    checkcols!(data, standardnames, standardtypes, bounds, "CPro")
-    new{T}(data)
-  end #constructor 1 CPro
-
-  """
-  Modified constructor of `CPro` reading data from hdf `files` for all given `sattime` indices
-  using MATLAB session `ms` and `lidarprofile` data, if data is above `altmin`.
-  """
-  function CPro{T}(
-    ms::mat.MSession,
-    files::Vector{String},
-    timeindex::Vector{UnitRange{Int}},
-    lidarprofile::NamedTuple,
-    saveobs::Union{String,Bool}="abs"
-  ) where T
-    # Return default empty struct if files are empty
-    (isempty(files) || saveobs === false || isempty(saveobs)) && return CPro{T}()
-    # Initialise arrays
-    # essential data
-    utc = Vector{Vector{DateTime}}(undef, length(files))
-    lat = Vector{Vector{T}}(undef, length(files))
-    lon = Vector{Vector{T}}(undef, length(files))
-    fcf = Vector{Vector{Vector{<:Union{Missing,UInt16}}}}(undef, length(files))
-    # non-essential data
-    ec532 = Vector{Vector{Vector{<:Union{Missing,T}}}}(undef, length(files))
-    Htropo = Vector{Vector{T}}(undef, length(files))
-    temp = Vector{Vector{Vector{<:Union{Missing,T}}}}(undef, length(files))
-    pres = Vector{Vector{Vector{<:Union{Missing,T}}}}(undef, length(files))
-    rH = Vector{Vector{Vector{<:Union{Missing,T}}}}(undef, length(files))
-    iwc = Vector{Vector{Vector{<:Union{Missing,T}}}}(undef, length(files))
-    deltap = Vector{Vector{Vector{<:Union{Missing,T}}}}(undef, length(files))
-    cad = Vector{Vector{Vector{<:Union{Missing,Int8}}}}(undef, length(files))
-    night = Vector{BitVector}(undef, length(files))
-    # Loop over files with cloud profile data
-    for (i, file) in enumerate(files)
-      ## Retrieve cloud profile data; assumes faulty files are filtered by SatData
-      mat.put_variable(ms, :file, file)
-      mat.eval_string(ms,
-        """
-        clear t longitude latitude
-        try
-          t = hdfread(file, 'Profile_UTC_Time');
-          t = t(:, 2);
-          longitude = hdfread(file, 'Longitude');
-          longitude = longitude(:, 2);
-          latitude = hdfread(file, 'Latitude');
-          latitude = latitude(:, 2);
-          FCF = hdfread(file, 'Atmospheric_Volume_Description');
-          EC532 = hdfread(file, 'Extinction_Coefficient_532');
-          Htropo = hdfread(file, 'Tropopause_Height');
-          temp = hdfread(file, 'Temperature');
-          pres = hdfread(file, 'Pressure');
-          rH = hdfread(file, 'Relative_Humidity');
-          IWC = hdfread(file, 'Ice_Water_Content_Profile');
-          deltap = hdfread(file, 'Particulate_Depolarization_Ratio_Profile_532');
-          CAD = hdfread(file, 'CAD_Score');
-          daynight = hdfread(file, 'Day_Night_Flag');
-        end
-
-        data = struct('time', t, 'lat', latitude, 'lon', longitude, ...
-          'FCF', FCF, 'EC532', EC532, 'Htropo', Htropo, 'temp', temp, 'pres', pres, ...
-          'rH', rH, 'IWC', IWC, 'deltap', deltap, 'CAD', CAD, 'daynight', daynight);
-        """
-      )
-      # Get data from MATLAB
-      data = mat.jdict(mat.get_mvariable(ms, :data))
-      # Convert time to UTC DateTime
-      utc[i] = convert_utc.(data["time"])
-      utc[i] = utc[i][timeindex[i]]
-      # Extract essential data
-      lat[i], lon[i] = data["lat"][timeindex[i]], data["lon"][timeindex[i]]
-      fcf[i] = get_lidarcolumn(UInt16, data["FCF"], lidarprofile, coarse=false)[timeindex[i]]
-      # Extract non-essential data
-      ec532[i] = get_lidarcolumn(T, data["EC532"], lidarprofile, missingvalues = -9999)[timeindex[i]]
-      Htropo[i] = 1000data["Htropo"][timeindex[i]]
-      temp[i] = get_lidarcolumn(T, data["temp"], lidarprofile, missingvalues = -9999)[timeindex[i]]
-      pres[i] = get_lidarcolumn(T, data["pres"], lidarprofile, missingvalues = -9999)[timeindex[i]]
-      rH[i] = get_lidarcolumn(T, data["rH"], lidarprofile, missingvalues = -9999)[timeindex[i]]
-      iwc[i] = get_lidarcolumn(T, data["IWC"], lidarprofile, missingvalues = -9999)[timeindex[i]]
-      deltap[i] = get_lidarcolumn(T, data["deltap"], lidarprofile, missingvalues = -9999)[timeindex[i]]
-      cad[i] = get_lidarcolumn(Int8, data["CAD"], lidarprofile, coarse=false,
-        missingvalues = -127)[timeindex[i]]
-      night[i] = Bool.(data["daynight"])[timeindex[i]]
-    end #loop over files
-
-    # Rearrange time vector and get time range
-    utc = [utc...;]
-    # Rearrange FCF vector and convert to symbols
-    fcf = [fcf...;]
-    avd =  Vector{Vector{Union{Missing,Symbol}}}(undef, length(fcf))
-    for i = 1:length(fcf)
-      vect = Vector{Union{Missing,Symbol}}(undef, length(fcf[i]))
-      for j = 1:length(fcf[i])
-        vect[j] = ismissing(fcf[i][j]) ? missing :
-          feature_classification(classification(fcf[i][j])...)
-      end
-      avd[i] = vect
+    #* Internal constructor with data checks
+    function SatData{T}(time::Vector{DateTime}, lat::Vector{<:AbstractFloat}, lon::Vector{<:AbstractFloat}) where T
+        length(time) == length(lat) == length(lon) ||
+            throw(DimensionMismatch("SatData vectors must have equal length"))
+        checklimits(time, DateTime(2000), Dates.now(), "time")
+        checklimits(lat, T(-90), T(90), "latitude")
+        checklimits(lon, T(-180), T(180), "longitude")
+        new{T}(time, lat, lon)
     end
-    # Construct and standardise data
-    data = DataFrame(time=utc, lat=[lat...;], lon=[lon...;],
-      atmos_state=avd, EC532=[ec532...;], Htropo = [Htropo...;],
-      temp=[temp...;], pressure = [pres...;], rH = [rH...;],
-      IWC = [iwc...;], deltap = [deltap...;],
-      CADscore = [cad...;], night = [night...;])
-    # Instantiate new struct
-    new{T}(data)
-  end #constructor 2 CPro
-end #struct CPro
-
-
-"""
-    CPro{T}() where T
-
-External constructor for empty `CPro` struct.
-"""
-CPro{T}() where T = CPro{T}(DataFrame(time = DateTime[], lat = T[], lon = T[],
-  atmos_state = Vector{Symbol}[], EC532 = Vector{T}[], Htropo = T[], temp = Vector{T}[],
-  pressure = Vector{T}[], rH = Vector{T}[], IWC = Vector{T}[],
-  deltap = Vector{T}[], CADscore = Vector{Int8}[], night = BitVector()))
-
-"""
-    CPro{T}(cpro::CPro) where T
-
-External `CPro` constructor for conversion of floating point precision.
-"""
-function CPro{T}(cpro::CPro) where T
-  convert_floats!(cpro.data, T)
-  CPro{T}(cpro.data)
 end
 
-"""
-    CPro(args...; kwargs...)
 
-Default CPro constructor for Float32.
+#* Main constructor for SatData from file
+function SatData{T}(file::AbstractString) where T<:AbstractFloat
+    # Open HDF5 file and read relevant datasets
+    t, lat, lon = h5.h5open(file, "r") do data
+        read(data, "Profile_UTC_Time")[2,:],
+        read(data, "Latitude")[2,:],
+        read(data, "Longitude")[2,:]
+    end
+    # Convert time to DateTime UTC
+    time = convert_utc.(t)
+    # CReturn SatData
+    return SatData{T}(time, lat, lon)
+end #constructor 2 SatData
+
+#* Constructor for empty SatData
+SatData{T}() where T<:AbstractFloat = SatData{T}(DateTime[], T[], T[])
+
+#* Constructor for default Float32 SatData
+SatData(args...; kwargs...) = SatData{Float32}(args...; kwargs...)
+
+#* Constructor for floating point type promotion
+SatData{T}(sat::SatData) where T<:AbstractFloat = SatData{T}(
+  sat.time, T.(sat.lat), T.(sat.lon)
+)
+
 """
-CPro(args...; kwargs...) = CPro{Float32}(args...; kwargs...)
+    SatTrack{T}(args...; kwargs...) where T<:AbstractFloat
+
+Alias constructor for [`SatData`](@ref). If `T` is omitted, `Float32` will be used by default.
+"""
+function SatTrack end
+
+SatTrack{T}(args...; kwargs...) where T<:AbstractFloat = SatData{T}(args...; kwargs...)
+SatTrack(args...; kwargs...) = SatData{Float32}(args...; kwargs...)
+
+
+"""
+# struct SatSet{T<:AbstractFloat} <: SecondarySet{T}
+
+Satellite data with fields
+- `granules::StructArray{SatData{T}}`
+- `metadata::SecondaryMetadata{T}`
+
+The `granules` contains data from individual satellite data files as `StructArray{SatData{T}}`
+with each `SatData{T}` storing continuous satellite track data from one file.
+
+see also [`SatTrack`](@ref), [`SatData`](@ref), and [`SecondaryMetadata`](@ref)
+
+## Instantiation
+
+    function SaSet{T}(
+      folders::String...;
+      type::Symbol=:undef,
+      attachments = nothing
+    ) where T<:AbstractFloat -> struct SatData
+
+Construct `SatSet{T}` from any number of absolute or relative folder paths given as string.
+SatData searches for `.h5` files in all folders recursively and determines the data type
+(`CLay` or `CPro`) from the majority of found files in the first valid folder automatically
+unless the type is specified with a Symbol `:CLay` or `:CPro` by the keyword argument `type`.
+Only one type of satellite data can be stored in `SatData`. All floating point data will be
+stored with the precision of `T`. If omitted, `Float32` will be used by default.
+
+Alternatively, handover all fields to the unmodified constructor, where basic data validity
+checks will be performed:
+
+    SatData{T}(granules::StructArray{SatData{T}}, metadata::SecondaryMetadata{T}) where T<:AbstractFloat
+
+Floating point type promotion is possible with:
+
+    SatSet{T}(sat::SatSet) where T<:AbstractFloat
+"""
+struct SatSet{T<:AbstractFloat} <: SecondarySet{T}
+    granules::StructArray{SatData{T}}
+    metadata::SecondaryMetadata{T}
+
+    #* Internal constructor with data checks
+    function SatSet{T}(
+        granules::StructArray{SatData{T}},
+        metadata::SecondaryMetadata{T}
+    ) where T<:AbstractFloat
+        # Check that granules and metadata lengths match
+        length(granules) == df.nrow(metadata.granules) ||
+            throw(DimensionMismatch("SatSet granules and metadata length mismatch"))
+        new{T}(granules, metadata)
+    end
+end
+
+
+#* Main constructor for SatSet
+function SatSet{T}(
+    folders::String...;
+    type::Symbol = :undef,
+    attachments = nothing
+)::SatSet{T} where T<:AbstractFloat
+    # Scan folders for satellite data files
+    t0 = Dates.now()
+    paths = []
+    for folder in folders
+        files = scandir(folder, ".h5")
+        type = sat_datafiles!(files, type)
+        push!(paths, (; root = folder, files))
+    end
+    # Loop over found files and load data
+    granules = StructArray{SatData{T}}(undef, 0)
+    metadata = DataFrame(file = String[], root = UInt16[], tstart = DateTime[], tstop = DateTime[],
+        latmin = T[], latmax = T[], elonmin = T[], elonmax = T[], wlonmin = T[], wlonmax = T[]
+    )
+    roots = ds.OrderedDict{String,UInt16}()
+    pm.@showprogress dt=1 desc="load sat data..." for data in paths
+        root = realpath(data.root)
+        haskey(roots, root) || (roots[root] = length(roots) + 1)
+        for file in data.files
+            datafile = joinpath(root, file)
+            try
+                granule = SatData{T}(datafile)
+                elonmin, elonmax = lonextrema(granule.lon, ≥)
+                wlonmin, wlonmax = lonextrema(granule.lon, <)
+                push!(metadata, (;file, root = roots[root],
+                    tstart = granule.time[1], tstop = granule.time[end],
+                    latmin = minimum(granule.lat), latmax = maximum(granule.lat),
+                    elonmin, elonmax, wlonmin, wlonmax
+                ))
+                push!(granules, granule)
+            catch e
+                @warn "read error; data skipped" file exception=(e, catch_backtrace())
+            end
+        end
+    end
+    # Return SatSet constructor
+    tend = Dates.now()
+    if isempty(metadata)
+        @warn "No satellite data files successfully loaded"
+        date = (start = Date(9999), stop = Date(9999))
+    else
+        date = (start = minimum(metadata.tstart), stop = maximum(metadata.tstop))
+    end
+    return SatSet{T}(granules, SecondaryMetadata{T}(
+        metadata,
+        ds.OrderedDict(v => k for (k,v) in roots),
+        type,
+        date,
+        tz.ZonedDateTime(tend, tz.localzone()),
+        tend - t0,
+        attachments
+    ))
+end
+
+#* Constructor for emtpy SatSet
+SatSet{T}() where T<:AbstractFloat = SatSet{T}(
+  StructArray{SatData{T}}(undef, 0),
+  SecondaryMetadata{T}()
+)
+
+#* Constructor for default Float32 SatSet
+SatSet(args...; kwargs...) = SatSet{Float32}(args...; kwargs...)
+
+#* Constructor for floating point type promotion
+SatSet{T}(sat::SatSet) where T<:AbstractFloat = SatSet{T}(
+  SatData{T}.(sat.granules), SecondaryMetadata{T}(sat.metadata)
+)
+
+
+## stucts for observations
